Move inbox state method into mod
