+++ bindings_node/src/inbox_state.rs
@@ -1,7 +1,9 @@
-use napi::bindgen_prelude::BigInt;
+use napi::bindgen_prelude::{BigInt, Result};
 use napi_derive::napi;
 use xmtp_cryptography::signature::ed25519_public_key_to_address;
 use xmtp_id::associations::{AssociationState, MemberIdentifier};
+
+use crate::{mls_client::NapiClient, ErrorWrapper};
 
 #[napi(object)]
 pub struct NapiInstallation {
@@ -35,5 +37,39 @@ impl From<AssociationState> for NapiInboxState {
         .collect(),
       account_addresses: state.account_addresses(),
     }
+  }
+}
+
+#[napi]
+impl NapiClient {
+  /**
+   * Get the client's inbox state.
+   *
+   * If `refresh_from_network` is true, the client will go to the network first to refresh the state.
+   * Otherwise, the state will be read from the local database.
+   */
+  #[napi]
+  pub async fn inbox_state(&self, refresh_from_network: bool) -> Result<NapiInboxState> {
+    let state = self
+      .inner_client()
+      .inbox_state(refresh_from_network)
+      .await
+      .map_err(ErrorWrapper::from)?;
+    Ok(state.into())
+  }
+
+  #[napi]
+  pub async fn get_latest_inbox_state(&self, inbox_id: String) -> Result<NapiInboxState> {
+    let conn = self
+      .inner_client()
+      .store()
+      .conn()
+      .map_err(ErrorWrapper::from)?;
+    let state = self
+      .inner_client()
+      .get_latest_association_state(&conn, &inbox_id)
+      .await
+      .map_err(ErrorWrapper::from)?;
+    Ok(state.into())
   }
 }
